requirements.txt
# Core dependencies (required)
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
aiofiles>=23.0.0

# Optional high-quality backend
# pip install vtracer
